Update issue 2722
This commit introduces a custom py2exe command, inspired by the InVEST
py2exe command.  Just before building the binaries, we zip up the
required and static data and put them into a zipfile.  I'm not
necessarily intending on keeping this, but it might be a decent way to
distribute a minimum set of data while still allowing us to install more
data later on.