Removing references to faulthandler.
Since it's in the adept package, it should be imported correctly by
pyinstaller.  The issue I was having was with pyinstaller not bringing
along the faulthandler EGG that I was expecting it to.  My solution was
to build an sdist for faulthandler and then install that to the
virtualenv.  This would build and install the C extension and just put
the necessary files into the site-packages dir, as I expect.