Update issue 3118
I think I found the issue with my imports ... I was setting a language
string BEFORE the variables LANG_ENGLISH and LANG_SPANISH were defined,
so both were set to the same value of 0.

By changing the import location to be AFTER the language dialog (which
I guess defines those couple of variables, if they aren't already
defined when I list the languages I wish to load), the language codes
are known.