Update issue 3362
Reverting the palisades substate to what it was for MAFE-T-1.0.3 so we
can get a temporary release out for cambodia.

This is because I'm currently in the middle of trying to fix a bunch of
localization issues when fetching input from the user and palisades is
in an unstable state.